requirements.txt
# ── Core ──────────────────────────────────────────────────────
pymupdf>=1.24.5
python-docx==1.1.0
cairosvg==2.7.1
Pillow>=10.4.0
docx2pdf>=0.1.8

# ── Telegram Bot ─────────────────────────────────────
python-telegram-bot>=21.0
python-dotenv>=1.0.0

# ── Telemetry ────────────────────────────────────────
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
grpcio>=1.62.1
